+++ serde_arrow/src/arrow2_impl/serialization.rs
@@ -133,6 +133,7 @@ fn build_array(builder: ArrayBuilder) -> Result<Box<dyn Array>> {
             build_array(*builder.element)?,
             build_validity(builder.validity),
         )?)),
+        A::FixedSizedList(_) => fail!("FixedSizedList is not supported by arrow2"),
         A::Struct(builder) => {
             let mut values = Vec::new();
             for (_, field) in builder.named_fields {