+++ serde_arrow/src/internal/schema/from_samples/mod.rs
@@ -27,6 +27,8 @@ impl Tracer {
         let mut tracer = Tracer::new(String::from("$"), String::from("$"), Arc::new(options));
         samples.serialize(OuterSequenceSerializer(&mut tracer))?;
         tracer.finish()?;
+        tracer.check()?;
+
         Ok(tracer)
     }
 }
@@ -234,8 +236,8 @@ impl<'a> serde::ser::Serializer for TracerSerializer<'a> {
         }
     }
 
-    fn serialize_bytes(self, _: &[u8]) -> Result<Self::Ok> {
-        fail!("cannot trace bytes")
+    fn serialize_bytes(self, val: &[u8]) -> Result<Self::Ok> {
+        val.serialize(self)
     }
 
     fn serialize_none(self) -> Result<Self::Ok> {