+++ serde_arrow/src/lib.rs
@@ -205,6 +205,7 @@ pub mod _impl {
                         ArrowPrimitiveType,
                         BooleanArray,
                         DictionaryArray,
+                        FixedSizeListArray,
                         GenericListArray,
                         GenericStringArray,
                         LargeStringArray,
@@ -296,7 +297,7 @@ pub mod _impl {
     // Reexport for tests
     #[doc(hidden)]
     pub use crate::internal::{
-        error::PanicOnError,
+        error::{PanicOnError, PanicOnErrorError},
         serialization::array_builder::ArrayBuilder,
     };
 }
@@ -393,4 +394,13 @@ pub mod schema {
     /// Renamed to [`SerdeArrowSchema`]
     #[deprecated = "serde_arrow::schema::Schema is deprecated. Use serde_arrow::schema::SerdeArrowSchema instead"]
     pub type Schema = SerdeArrowSchema;
+
+    /// Support for [canonical extension types][ext-docs]. This module is experimental without semver guarantees.
+    ///
+    /// [ext-docs]: https://arrow.apache.org/docs/format/CanonicalExtensions.html
+    pub mod ext {
+        pub use crate::internal::schema::extensions::{
+            FixedShapeTensorField, VariableShapeTensorField,
+        };
+    }
 }