+++ Changes.md
@@ -2,6 +2,7 @@
 
 ## 0.11.6
 
+- Add support for the `FixedSizeList(n)` data type (only `arrow`)
 - Add support to overwrite field definitions with `TracingOptions::overwrite`
 - Add support to serialize enums without data (e.g., `enum E { A, B, C}`) as
   strings by setting the corresponding field to a string value (`Utf`,